requirements.txt
torch==2.1.0
transformers==4.37.2
tokenizers==0.15.1
pandas==2.1.4
numpy==1.24.3
tqdm==4.66.1
pillow==10.1.0
sentence-transformers==2.2.2
accelerate==0.21.0
bitsandbytes==0.41.1
huggingface-hub==0.19.4 